+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.3.0.0
 * Split machine-state and opcodes into separate modules.
+* Add `Intcode.Parse`
+* Add `Intcode.hRunIO`
 
 ## 0.2.0.0
 * Rename 'intCodeToList' to 'intcodeToList'